Fixes #8: Possible race condition preventing some roundHashes resets
I am not 100% sure that this addresses the problem, since it has not
been reliably reproduced.  This is my best guess, that the roundHashes
reset during storing a mined block is called in the middle of a
coincidental storeBulkShares call, causing the latter to be unaware of
the intent to reset and carry forward roundHashes.

storeBulkShares now recognizes a found block and resets roundHashes.
This change *potentially* deprecates the roundHashes reset in the block
mined logic, but I'm going to leave it for now.