Force a template refresh during long network mining gaps
When the network takes a long time to discover its next block, the
timestamp within the blocktemplate that miners are churning can be
very stale.  Further, new transactions aren't being sucked in.

This forces a template refresh if the blockTargetTime for the
configured coin passes without the network finding a block.